+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### v4.20.0 Features
 
+- `[Datagrid]` Added a Sort Function to the datagrid column to allow the value to be formatted for the sort. ([#1766](https://github.com/infor-design/enterprise/issues/2274)))
+
 ### v4.20.0 Fixes
 
 - `[Datagrid]` Fixes tooltip on the non displayed table errors. ([#2264](https://github.com/infor-design/enterprise/issues/2264))